Simplify sparql querying
The current state of querying is weirdly complex, and there seem to be a number of interfaces which do very little (or, even, add very little value).
This seems, in part, because the querier wants to allow for many different types fo SPARQL endpoint, which all have their own foibles. But... that seems daft. The Repo querier is literally just a way of wrapping http requests, and parsing some application/sparql-results+json responses. And even that gets done in another function. So why bother at all? Instead, we make Repo a generic querier. It does a GET or a POST, and then returns some stuff. It takes a string. Let anybody else who requires a specialised client bring their own. The important bits here are the parser, and the query bank.