Shift repo into subpackage
The repo code, while very useful, is pretty much just an http client which calls the important part of the sparql parser. There is no real reason why all users need to pull this in when all they want is a sparql parser, and so we move it out and allow developers to either use it, or provide their own.